move dos exe launching to winoldap and make the dos emulator to use configurable.  also fix launching files with shellexecute16 that have spaces in filenames
