+++ otvdm.ini
@@ -51,3 +51,8 @@
 
 ; Don't report running under Windows NT.
 ;IAmNotNT=1
+;
+; Set application to launch MSDOS programs with
+; Works with msdos player, dosbox requires a config file to mount virtual drives
+; (default: otvdm.exe)
+;dosemu=msdos.exe
